PC-24819 break formatters into formatters.go
